test DPT validate_payload
